Format numeric inputs on blur.
To ensure consistent formatting, gallon and mileage inputs are formatted
using the en-US locale, with varying precision. Additionally, when
switching between odometer mileage and trip mileage input modes, the
other value is retained in case the user switches back. This
necessitated keeping both values in the state, and choosing which is
displayed depending on mode.

There is an issue where attempting to format an already formatted number does not
return the same number, e.x. 200,000 becomes 200.0. This is because
`parseFloat` and the `Number` constructor treat the comma separator as a
decimal indicator. To work around this, all commas are striped from the
input string. Similarly, commas are not allowed in `<input type="number">`,
so numeric inputs are now `<input type="text">`.