Implement `parseNumber`, inverse of `formatNumber`.
After setting the value of inputs to numeric strings formatted by `formatNumber`,
functions expecting to convert the value to a number using the `Number`
constructor or `parseFloat` choke if the numeric string contains a comma
delimiter. `parseNumber` removes the comma delimiter, effectively
undoing `formatNumber`. Also made `formatNumber` return an empty string
on NaN; an empty string clears the input, which is the expected behavior
for invalid input when blurred.