Add `NumericInput` form component.
This component integrates the blur formatting logic into the
input, so that reducers do not have to handle blur separately.
The component only exposes a `setValue` which passes up the
value on change, and passes up the formatted value on blur;
this is to consolidate changes through a single interface.